requirements.txt
alembic==1.8.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
flask-marshmallow==0.15.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
greenlet>=1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
pathspec==0.9.0
psycopg2==2.9.3
PyYAML==6.0
requests==2.28.1
SQLAlchemy==2.0.15
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
python-decouple==3.8
aiohttp==3.8.4